// hub/frontend/lib/api/generated/model/eventDeliveryRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.24
 */
import type { EventDeliveryEventSummary } from './eventDeliveryEventSummary';
import type { EventDeliveryReadRequestHeaders } from './eventDeliveryReadRequestHeaders';

export interface EventDeliveryRead {
  attemptCount: number;
  createdAt: string;
  destinationType: string;
  destinationUrlRedacted: string;
  errorMessage: string;
  event?: EventDeliveryEventSummary | null;
  eventRecordId: string;
  eventRuleId: string | null;
  id: string;
  lastAttemptAt: string | null;
  nextAttemptAt: string | null;
  requestHeaders: EventDeliveryReadRequestHeaders;
  responseBody: string;
  responseStatus: number | null;
  status: string;
  updatedAt: string;
}